+++ Server/routes/userRoutes.js
@@ -6,6 +6,7 @@ const authMiddleware = require("../middleware/authMiddleware");
 // Define the routes
 router.post("/register", userController.registerUser);
 router.post("/login", userController.loginUser);
+router.post("/logout", authMiddleware, userController.logoutUser);
 router.get("/", userController.getAllUsers);
 router.get("/:userId", userController.getUserById);
 router.put("/:userId", authMiddleware, userController.updateUser);